requirements.txt
certifi==2019.9.11
chardet==3.0.4
construct==2.5.3
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-appconf==1.0.3
django-ckeditor==5.7.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==1.1.0
future==0.17.1
haystack==0.42
idna==2.8
jsonfield==2.0.2
Markdown==3.1.1
oauthlib==3.1.0
pefile==2019.4.18
pilkit==2.0
Pillow==6.1.0
psycopg2==2.8.3
Pygments==2.4.2
python-ptrace==0.9.4
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
uWSGI==2.0.18
